move take while serialization tests to kernel tests
