+++ tensorflow/python/data/experimental/kernel_tests/take_while_test.py
@@ -21,6 +21,7 @@ from absl.testing import parameterized
 import numpy as np
 
 from tensorflow.python.data.experimental.ops import take_while_ops
+from tensorflow.python.data.kernel_tests import checkpoint_test_base
 from tensorflow.python.data.kernel_tests import test_base
 from tensorflow.python.data.ops import dataset_ops
 from tensorflow.python.framework import combinations
@@ -124,6 +125,21 @@ class TakeWhileTest(test_base.DatasetTestBase, parameterized.TestCase):
     dataset = dataset_ops.Dataset.range(10).apply(
         take_while_ops.take_while(lambda x: x < 2)).repeat(5)
     self.assertDatasetProduces(dataset, np.tile([0, 1], 5))
+
+class TakeWhileDatasetCheckpointTest(checkpoint_test_base.CheckpointTestBase,
+                                     parameterized.TestCase):
+
+  def _build_dataset(self, num_elements, upper_bound):
+    return dataset_ops.Dataset.range(num_elements).apply(
+        take_while_ops.take_while(lambda x: x < upper_bound))
+
+  @combinations.generate(
+      combinations.times(
+          test_base.default_test_combinations(),
+          combinations.combine(num_elements=[10, 23], upper_bound=[10, 23])))
+  def testCore(self, num_elements, upper_bound):
+    self.run_core_tests(lambda: self._build_dataset(num_elements, upper_bound),
+                        min(num_elements, upper_bound))
 
 
 if __name__ == "__main__":
